Make xinitrc-common survive set -u
What is set -u? set -u makes expanding an undefined variable an error,
which (here) would happen when XKB_IN_USE != yes. This seems like a
strange thing to want, and a rather stern thing to impose upon the shell
scripts of others, but whatever let's just write the test a different
way.

Signed-off-by: Adam Jackson <ajax@redhat.com>